Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 124ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240601/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -7,7 +7,7 @@ public class Main {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
 
         long startNumber = 1; // 범위의 시작
-        long endNumber = Integer.parseInt(reader.readLine()); // 컴퓨터가 선택한 수 (범위의 끝)
+        long endNumber = Long.parseLong(reader.readLine()); // 컴퓨터가 선택한 수 (범위의 끝)
         
         StringTokenizer tokenizer = new StringTokenizer(reader.readLine());
         long startTarget = Long.parseLong(tokenizer.nextToken()); // 목표 숫자 범위의 시작
